requirements.txt
absl-py==2.0.0
accelerate==0.29.3
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
bleach==6.1.0
blessed==1.20.0
blinker==1.4
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloud-init==23.4.4
cloudpathlib==0.16.0
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
confection==0.1.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.0
cryptography==3.4.8
cupy-cuda12x==12.3.0
cupy-wheel==12.3.0
cycler==0.12.1
cymem==2.0.8
cytoolz==0.12.2
datasets==2.16.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
dill==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.4.2
docopt==0.6.2
ec2-hibinit-agent==1.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastrlock==0.8.2
filelock==3.13.1
flatbuffers==23.5.26
floret==0.10.5
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gpustat==1.1.1
grpcio==1.60.0
h5py==3.10.0
hibagent==1.0.1
httplib2==0.20.2
huggingface-hub==0.20.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
install==1.3.5
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jeepney==0.7.1
jellyfish==1.0.3
Jinja2==3.0.3
joblib==1.3.2
jsonpatch==1.32
jsonpickle==3.0.2
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kaggle==1.6.12
keras==2.15.0
keras-tuner==1.4.7
keyring==23.5.0
kiwisolver==1.4.5
kt-legacy==1.0.5
langcodes==3.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
leveldb==0.201
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mlp-mixer==0.0.1
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
murmurhash==1.0.10
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pbr==5.8.0
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.0
pydotplus==2.0.2
Pygments==2.17.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymongo==4.6.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyphen==0.14.0
pyrsistent==0.18.1
pyserial==3.5
pyspellchecker==0.7.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
python-slugify==8.0.4
pytz==2022.1
PyYAML==5.4.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rouge_score==0.1.2
rsa==4.9
sacred==0.8.5
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
screen-resolution-extra==0.0.0
seaborn==0.13.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sos==4.5.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.11
stack-data==0.6.3
streamlit==1.33.0
sympy==1.12
systemd-python==234
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-addons==0.23.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
testresources==2.0.1
text-unidecode==1.3
textacy==0.13.0
thinc==8.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.36.2
triton==2.1.0
Twisted==22.1.0
typeguard==2.13.3
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.26.5
wadllib==1.3.6
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.12
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
xkit==0.0.0
xxhash==3.4.1
yarl==1.9.4
zipp==1.0.0
zope.interface==5.4.0
